ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
scikit-learn==1.6.1
nltk==3.9.1
pandas==2.2.3
numpy==2.2.1
python-dotenv==1.0.1
joblib==1.4.2
pydantic==2.10.4
